Move packages and use serviceloader
