+++ shims/spark30/src/main/resources/META-INF/services/com.nvidia.spark.rapids.SparkShimLoader
@@ -1,0 +1,1 @@
+com.nvidia.spark.rapids.shims.spark30.Spark30ShimLoader
