+++ shims/spark31/src/main/scala/com/nvidia/spark/rapids/shims/spark31/GpuSortMergeJoinExec.scala
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims
+package com.nvidia.spark.rapids.shims.spark31
 
 import com.nvidia.spark.rapids._
 import com.nvidia.spark.rapids.GpuMetricNames._
 
-//import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, BuildSide}
-import org.apache.spark.sql.execution.joins.{BuildLeft, BuildRight, BuildSide}
+import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, BuildSide}
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.execution.joins.BroadcastNestedLoopJoinExec
 import org.apache.spark.sql.execution.joins.ShuffledHashJoinExec
@@ -38,7 +37,7 @@ import org.apache.spark.internal.Logging
 
 
 
-class GpuSortMergeJoinMeta30(
+class GpuSortMergeJoinMeta(
     join: SortMergeJoinExec,
     conf: RapidsConf,
     parent: Option[RapidsMeta[_, _, _]],
@@ -55,7 +54,7 @@ class GpuSortMergeJoinMeta30(
   override def tagPlanForGpu(): Unit = {
 
     // Use conditions from Hash Join
-    GpuHashJoin30.tagJoin(this, join.joinType, join.leftKeys, join.rightKeys, join.condition)
+    GpuHashJoin.tagJoin(this, join.joinType, join.leftKeys, join.rightKeys, join.condition)
 
     if (!conf.enableReplaceSortMergeJoin) {
       willNotWorkOnGpu(s"Not replacing sort merge join with hash join, " +
@@ -80,7 +79,7 @@ class GpuSortMergeJoinMeta30(
     }
   }
   override def convertToGpu(): GpuExec = {
-    GpuShuffledHashJoinExec30(
+    GpuShuffledHashJoinExec(
       leftKeys.map(_.convertToGpu()),
       rightKeys.map(_.convertToGpu()),
       join.joinType,
