+++ tests/pom.xml
@@ -30,30 +30,30 @@
     <description>RAPIDS plugin for Apache Spark integration tests</description>
     <version>0.2.0-SNAPSHOT</version>
 
+    <properties>
+        <slf4j.version>1.7.30</slf4j.version>
+    </properties>
     <profiles>
         <profile>
             <id>spark31</id>
-            <dependencies>
-            </dependencies>
+            <properties>
+                <spark.version>3.1.0</spark.version>
+            </properties>
         </profile>
     </profiles>
 
-	    <properties>
-	        <spark.version>3.0.0</spark.version>
-                <slf4j.version>1.7.30</slf4j.version>
-	    </properties>
     <dependencies>
-              <dependency>
-                <groupId>org.slf4j</groupId>
-                <artifactId>jul-to-slf4j</artifactId>
-                <version>${slf4j.version}</version>
-              </dependency>
-              <dependency>
-                <groupId>org.slf4j</groupId>
-                <artifactId>jcl-over-slf4j</artifactId>
-                <version>${slf4j.version}</version>
-                <!-- runtime scope is appropriate, but causes SBT build problems -->
-              </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>jul-to-slf4j</artifactId>
+            <version>${slf4j.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>jcl-over-slf4j</artifactId>
+            <version>${slf4j.version}</version>
+            <!-- runtime scope is appropriate, but causes SBT build problems -->
+        </dependency>
         <dependency>
             <groupId>org.scala-lang</groupId>
             <artifactId>scala-library</artifactId>
